% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction_wrapper.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Dimensional reduction of expression counts using 0-1 scaled expression levels, PCA, or LDA}
\usage{
dim_reduction(
  data_input,
  method = c("scale", "pca", "lda"),
  threshold = 0.5,
  lda_threshold = 0.8,
  focus = "group",
  lda_focus = "group"
)
}
\arguments{
\item{data_input}{Expression data, sample information, and biological process list uploaded using import_data()}
}
\description{
For each category of variables (e.g., group of genes related to a given biological process), reduces the dimensionality across variables to a single value for each sample.
}
\details{
If using method="scale", the function returns the average value of the 0-1 scaled variable within each category for each sample

If using method="pca", the function uses a Principal Component Analysis (PCA) on the expression matrix to derive the coordinates of the samples for each group. The top PC dimensions are retained based on a variance threshold that can be modified in the function (default: threshold = 0.5). The axis which maximises the variance in sample-averaged PCA coordinates for the user-defined focus (default: focus = "group") is defined. Each sample is projected onto this axis. The most extreme samples are set to have values of 0 and 1, with lower projection values corresponding to the side of the axis showing lower expression values.

If using method="lda", the function uses a Principal Component Analysis (PCA) on the expression matrix to derive the coordinates of the samples for each group. The top PC dimensions are retained based on a variance threshold that can be modified in the function (default: threshold = 0.5). A Linear Discriminant Analysis (LDA) is used to increase the separation between samplles based on a user-defined lda-focus (default: lda_focus = "group"). The top LD dimensions are retained based on a variance threshold that can be modified in the function (default: lda_threshold = 0.8). The axis which maximises the variance in sample-averaged LDA coordinates for the user-defined focus (default: focus = "group") is defined. Each sample is projected onto this axis. The most extreme samples are set to have values of 0 and 1, with lower projection values corresponding to the side of the axis showing lower expression values.
}
