% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction_pca_method.R
\name{pca_method}
\alias{pca_method}
\title{Dimensional reduction based on PCAs.}
\usage{
pca_method(data_input, pca_threshold, focus, correlation_method, pca_scale)
}
\arguments{
\item{data_input}{Counts data, sample information, and biological process list uploaded using import_data()}

\item{pca_threshold}{threshold of cumulative variance used to identify the number of PCs to keep,}

\item{focus}{grouping of samples used to determine the main axis of variance (default = "group")}

\item{correlation_method}{statistical test used to determine linear correlation between values extracted for each sample and mean scaled counts in the category (default = "spearman").}
}
\value{
A list with:
\itemize{
\item projection: data frame with sample, biological category, furthest sample groups based on the top PC dimensions retained, distance of each sample along a multidimensional line linking the most extreme samples, 0-1 scaled distance with direction based on counts of most extreme groups.
\item information: data frame with category, number of features (genes/proteins/others), method used ("PCA"), number of PCs kept, sum of variance kept, correlation between values extracted for samples in that biological category and their counts.
\item dimred_information: data frame with information needed for plot_dimensions(): biological category, number of PCs retained, percentage of variance described by PC1 and PC2, centroid, and slope of main axis of variance.
}
}
\description{
To be used as part of dim_reduction().
For each biological category, extraction of a single value per sample based on top dimensions from Principal Component Analyses
}
\details{
Perform PCA on the counts matrix of genes/proteins/others from a biological category.
identify the number of PCs corresponding to a user-defined threshold of variance,
if PC1 meets the threshold, extract the PC1 coordinate of each sample.
If not, identify the main axis of variance between the average PC coordinates of each group (or other user-defined sample grouping scheme),
project the samples onto this axis in the multidimensional space,
determine which side of the axis is associated with lower values based on average gene expression in 2 most extreme groups,
scale sample values from 0 to 1, setting 0 to be the on the side of lower values.
}
