% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot boxplots for data inspection}
\usage{
plot_boxplot(
  data_input,
  dim_reduction_output,
  stat_choice = "nonparametric",
  pvalue_adjustment = "bh",
  pvalue_threshold = 0.05,
  boxplot_focus = "group",
  boxplot_order = NULL,
  colour_palette = c(wes_palette("AsteroidCity1"), wes_palette("Chevalier1"),
    wes_palette("Darjeeling2"), wes_palette("Darjeeling1"),
    wes_palette("GrandBudapest1"), wes_palette("Moonrise3"), wes_palette("Moonrise2"))
)
}
\arguments{
\item{data_input}{Counts data, sample information, and biological feature list uploaded using import_data()}

\item{dim_reduction_output}{Output from dim_reduction()}

\item{stat_choice}{Choice of "nonparametric" (Kruskal + Dunn), or "parametric" (ANOVA + Tukey's HSD test). Default = "nonparametric".}

\item{pvalue_adjustment}{Choice of p-value adjustment methods for non-parametric testing. Same options as for dunn.test() from library(dunn.test). Default = "bh" (Benjamini-Hochberg)}

\item{pvalue_threshold}{Choice of p-value significance threshold (default = 0.05)}

\item{boxplot_order}{Manual modification of the order of the groups of boxplots. Must be provided as a vector, e.g., boxplot_order = c("s1","s2",...).}

\item{colour_palette}{Define colour palette (default: a mix of palettes from the "wesanderson" package).}

\item{boxplot_grouping}{Choice of sample grouping for boxplot visualisation and statistical testing. Default = "group" from data_input$sample_meta.}
}
\value{
A list of plots displaying the values yielded by dim_reduction() for each sample in each biological category. One plot per category.
}
\description{
Produce boxplots of the values extracted for each sample across categories, grouping samples based on default = "group" or user defined criteria.
}
