% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimensions.R
\name{plot_dimensions}
\alias{plot_dimensions}
\title{Print PCA/LDA plots and if <3 dimensions are retained, display main axis onto which each sample is projected.}
\usage{
plot_dimensions(data_input, dim_reduction_output, colour = "group", shape = NA)
}
\arguments{
\item{data_input}{Expression data, sample information, and gene list uploaded using import_data()}

\item{dim_reduction_output}{Output from dim_reduction()}

\item{colour}{Define colour of plotted data points based on column name from the sample information matrix (default = "group")}

\item{shape}{Define shape of plotted data points based on column name from the sample information matrix (default = NA)}
}
\value{
A list of plots displaying the coordinates of each sample for each biological category
}
\description{
Print PCA/LDA plots and if <3 dimensions are retained, display main axis onto which each sample is projected.
}
