% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_radar}
\alias{plot_radar}
\title{Plot radar charts following dim_reduction()}
\usage{
plot_radar(
  data_input,
  dim_reduction_output,
  category_list = NULL,
  colour = NULL
)
}
\arguments{
\item{data_input}{Expression data, sample information, and gene list uploaded using import_data()}

\item{dim_reduction_output}{Output from dim_reduction()}

\item{category_list}{Optional data frame with columns: 1) category names, 2) order on the radar plot. Default will follow the order of gene categories from input file used in import_data(). You do not have to include all categories originally provided, and you can play around with the order in which categories are displayed to manually highlight relationships between biological processes.}

\item{colour}{Optional colour scheme choice for different groups of samples. For optimal visualisation quality, we recommend exporting the outputs as PDFs and manually editing the colour scheme and figure design with vector editing software.}
}
\value{
A list of `ggradar` plots, one per group of samples.
}
\description{
Plot radar charts following dim_reduction()
}
\examples{
plot_list=plot_radar(data_input,dim_reduction_output,category_list=my_category_list)
}
