% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groups.R
\name{plot_radar_groups}
\alias{plot_radar_groups}
\title{Plot radar charts from PCA dim_reduction_output}
\usage{
plot_radar_groups(data_input, dim_reduction_output, category_list = NULL)
}
\arguments{
\item{data_input}{Expression data, sample information, and gene list uploaded using import_data()}

\item{dim_reduction_output}{Output from dim_reduction()}

\item{category_list}{Optional data frame with columns: 1) category names,2) order on the radar plot. Default will follow the order of gene categories from input file used in import_data()}
}
\value{
A list of `ggradar` plots,one per group
}
\description{
Plot radar charts from PCA dim_reduction_output
}
\examples{
plot_list=plot_radar(data_input,dim_reduction_output,category_list=my_category_list)
}
