% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction_wrapper.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Dimensional reduction}
\usage{
dim_reduction(
  data_input,
  method = c("scale", "pca", "lda"),
  pca_threshold = 0.5,
  lda_threshold = 0.8,
  pca_scale = TRUE,
  focus = "group",
  lda_focus = "group",
  correlation_method = "spearman",
  remove_effect_from = NULL
)
}
\arguments{
\item{data_input}{Counts data, sample information, and biological process list uploaded using import_data()}

\item{method}{Choice of scale, pca, or lda}

\item{pca_threshold}{when using method = "pca" or "lda", threshold of cumulative variance used to identify the number of PCs to keep,}

\item{lda_threshold}{when using method = "lda", threshold of cumulative variance used to identify the number of LDs to keep,}

\item{pca_scale}{parameter used inside of prcomp() function. Default = TRUE.}

\item{focus}{when using method = "pca" or "lda", grouping of samples used to determine the main axis of variance (default = "group")}

\item{lda_focus}{when using method = "lda", grouping of samples used to perform LDA analysis (default = "group")}

\item{correlation_method}{statistical test used to determine linear correlation between values extracted for each sample and mean scaled counts in the category (default = "spearman").}

\item{remove_effect_from}{use limma to remove the effect of one or a few factors on the counts matrix before analysis (default: none. Use with "treatment" or c("treatment1","treatment2)).}
}
\value{
When using method = "scale", the function returns a list with:
\itemize{
\item projection: data frame with sample, normalised_distance (normalized avg expr), category, furthest_groups.
\item information: data frame with number of features per category.
When using method = "pca", the function returns a list with:
\item projection: data frame with sample, biological category, furthest sample groups based on the top PC dimensions retained, distance of each sample along a multidimensional line linking the most extreme samples, 0-1 scaled distance with direction based on counts of most extreme groups.
\item information: data frame with category, number of features (genes/proteins/others), method used ("PCA"), number of PCs kept, sum of variance kept, correlation between values extracted for samples in that biological category and their counts.
\item dimred_information: data frame with information needed for plot_dimensions(): biological category, number of PCs retained, percentage of variance described by PC1 and PC2, centroid, and slope of main axis of variance.
When using method = "lda", the function returns a list with:
\item projection: data frame with sample, biological category, furthest sample groups based on the top PC dimensions retained, distance of each sample along a multidimensional line linking the most extreme samples, 0-1 scaled distance with direction based on counts of most extreme groups.
\item information: data frame with category, number of features (genes/proteins/others), method used ("PCA" or "LDA"), number of PCs and LDs kept, sum of variance kept, correlation between values extracted for samples in that biological category and their counts.
\item dimred_information: data frame with information needed for plot_dimensions(): biological category, number of PCs retained, percentage of variance described by PC1, PC2, LD1, and LD2, centroid, and slope of main axis of variance.
}
}
\description{
Dimensional reduction of counts using 0-1 scaling, PCA, or LDA. Refer to specific documentation for each reduction approach (?scale_method, ?pca_method, ?lda_method) for more.
}
\details{
For each biological category (e.g., group of genes related to a given biological process), reduces the dimensionality across features to a single value for each sample.

If using method="scale", the function returns the average value of the 0-1 scaled features within each category for each sample

If using method="pca", the function uses a Principal Component Analysis (PCA) on the counts matrix to derive the coordinates of the samples for each group. The top PC dimensions are retained based on a variance threshold (default: threshold = 0.5). The axis which maximises the variance in sample-averaged PCA coordinates for the user-defined focus (default: focus = "group") is defined. Each sample is projected onto this axis. The most extreme samples are set to have values of 0 and 1, with lower projection values corresponding to the side of the axis showing lower counts values.

If using method="lda", the function uses a Principal Component Analysis (PCA) on the counts matrix to derive the coordinates of the samples for each group. The top PC dimensions are retained based on a variance threshold (default: threshold = 0.5). When more than 1 PC is needed to reach threshold, a Linear Discriminant Analysis (LDA) is performed to increase the separation between samples based on a user-defined lda_focus (default: lda_focus = "group"). The top LD dimensions are retained based on a variance lda_threshold (default: lda_threshold = 0.8). The axis which maximises the variance in sample-averaged LDA coordinates for the user-defined focus (default: focus = "group") is defined. Each sample is projected onto this axis. The most extreme samples are set to have values of 0 and 1, with lower projection values corresponding to the side of the axis showing lower counts values.
}
