% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import sequencing data, sample metadata, and molecular feature list}
\usage{
import_data(counts_path, sample_meta_path, feature_meta_path)
}
\arguments{
\item{counts_path}{File path to counts matrix (molecular features as rownames, samples as column names, cell fill as counts level). This counts matrix should be normalised (e.g., variance-stabilised).}

\item{sample_meta_path}{File path to sample metadata (one row per sample. column names of this table should include "sample", "treatment". within the sample column, the names should match the column names of the counts matrix)}

\item{feature_meta_path}{File path to molecular feature (e.g., gene) metadata (one row per feature. column names of this table should include "feature", "category". within the molecular feature column, the names should feature within the row names of the counts matrix)}
}
\value{
A named list with elements: \code{counts}, \code{sample_meta}, and \code{feature_meta}.
}
\description{
Import sequencing data, sample metadata, and molecular feature list
}
