% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction_lda_method.R
\name{lda_method}
\alias{lda_method}
\title{For each gene category,calculation of distance between samples based on top dimensions from Principal Component Analyses}
\usage{
lda_method(data_input, threshold, lda_threshold, focus, lda_focus)
}
\arguments{
\item{threshold}{threshold of cumulative variance used to identify the number of PCs to keep.}

\item{expr}{Gene expression matrix (genes as rownames,samples as column names,cell fill as gene expression level). This gene expression matrix should be normalised (e.g.,variance-stabilised).}

\item{sample_meta}{Sample metadata (one row per sample. column names of this table should include "sample","group". within the sample column,the names should match the column names of the expression matrix)}

\item{gene_meta}{Gene metadata (one row per gene. column names of this table should include "gene","category". within the gene column,the names should feature within the row names of the expression matrix)}
}
\value{
A list with:
  - projection: data frame with distance of each sample along a multidimensional line linking the most extreme samples,category,and furthest group pair.
  - information: data frame with category,number of PCs kept,sum of variance kept,and number of genes considered. Distances obtained for the PCA
}
\description{
Perform PCA on the expression level of genes in a given gene category,
identify the number of PCs corresponding to a certain threshold of variance,
identify the axis of maximum variance between the average coordinates of each group,
project the samples onto this axis in the multidimensional space,
determine which side of the axis is associated with lower values based on average gene expression in 2 most extreme groups.
}
