% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Plot radars}
\usage{
plot_radar(
  data_input,
  dim_reduction_output,
  radar_grouping = "group",
  category_list = NA,
  colour_sample = NULL,
  colour_average = NULL,
  axis_label_size = 1,
  radar_label_size = 4,
  radar_label_position = "centre",
  width = 2,
  height = 2
)
}
\arguments{
\item{data_input}{Expression data, sample information, and gene list uploaded using import_data()}

\item{dim_reduction_output}{Output from dim_reduction()}

\item{radar_grouping}{Define how to group samples together on the radars. Default = "group" from sample information.}

\item{category_list}{Optional data frame with columns: 1) category names, 2) order on the radar plot. Default will follow the order of gene categories from input file used in import_data(). You do not have to include all categories originally provided in import_data(), and you can play around with the order in which categories are displayed to manually highlight relationships between biological processes.}

\item{colour_sample}{Optional colour scheme choice for the radar of each sample. This must be a dataframe with one "sample" column and one "colour" column. For optimal visualisation quality, we recommend exporting the outputs as PDFs and manually editing the colour scheme and figure design with vector editing software.}

\item{colour_average}{Optional colour scheme choice for the average radar of each group of samples. This must be a dataframe with one column corresponding to the sample grouping (name of column must match the argument "radar_grouping" - default = "group") and one "colour" column. For optimal visualisation quality, we recommend exporting the outputs as PDFs and manually editing the colour scheme and figure design with vector editing software.}

\item{axis_label_size}{Modify size of axis label}

\item{radar_label_size}{Modify size of radar label / title}

\item{radar_label_position}{Define whether the radar label is placed in the centre of the plot (default) or at the top.}

\item{width}{Define the horizontal spacing of the plots. Can help to better print the name of the different biological categories around the plot when they are long.}

\item{height}{Define the vertical spacing of the plots.}
}
\value{
A list of \code{ggradar} plots, one per group of samples.
}
\description{
Plot radar charts following dim_reduction()
}
\examples{
plot_list=plot_radar(data_input,dim_reduction_output,category_list=my_category_list)
}
